Conway's game of life in java using GUI
http://en.wikipedia.org/wiki/Conway's_Game_of_Life , When the program is
compiled then the program will ask the user to enter the number of rows
and columns they need in the program, then it asks if they want to
randomise the frame or not if pressed y(yes) then it puts green cells
all over the grid randomly if pressed n(no) then it prints the cell
design stored in the text file. NOTE :- the text file is not provided in
this project. Press start and the process starts the number of
generations is shown below on the right bottom screen. Press stop to
stop the code, press step to compute the next step. Note: some of the
code in this can be written in a simpler way.